Randomize background color for single event page header
